Follow up seperate out the type structs
